+++ wxc_sdk/as_api.py
@@ -68,9 +68,9 @@ __all__ = ['AsAccessCodesApi', 'AsAdminAuditEventsApi', 'AsAgentCallerIdApi', 'A
            'AsMeetingParticipantsApi', 'AsMeetingPreferencesApi', 'AsMeetingQandAApi', 'AsMeetingQualitiesApi',
            'AsMeetingTranscriptsApi', 'AsMeetingsApi', 'AsMembershipApi', 'AsMessagesApi', 'AsMonitoringApi',
            'AsMoveUsersJobsApi', 'AsMusicOnHoldApi', 'AsNumbersApi', 'AsOrganisationAccessCodesApi',
-           'AsOrganisationVoicemailSettingsAPI', 'AsOrganizationApi', 'AsOutgoingPermissionsApi', 'AsPagingApi',
-           'AsPeopleApi', 'AsPersonForwardingApi', 'AsPersonSettingsApi', 'AsPersonSettingsApiChild', 'AsPlayListApi',
-           'AsPreferredAnswerApi', 'AsPremisePstnApi', 'AsPriorityAlertApi', 'AsPrivacyApi',
+           'AsOrganisationVoicemailSettingsAPI', 'AsOrganizationApi', 'AsOutgoingPermissionsApi', 'AsPSTNApi',
+           'AsPagingApi', 'AsPeopleApi', 'AsPersonForwardingApi', 'AsPersonSettingsApi', 'AsPersonSettingsApiChild',
+           'AsPlayListApi', 'AsPreferredAnswerApi', 'AsPremisePstnApi', 'AsPriorityAlertApi', 'AsPrivacyApi',
            'AsPrivateNetworkConnectApi', 'AsPushToTalkApi', 'AsRebuildPhonesJobsApi', 'AsReceptionistApi',
            'AsReceptionistContactsDirectoryApi', 'AsRecordingsApi', 'AsReportsApi', 'AsRestSession', 'AsRolesApi',
            'AsRoomTabsApi', 'AsRoomsApi', 'AsRouteGroupApi', 'AsRouteListApi', 'AsSCIM2BulkApi', 'AsSCIM2GroupsApi',
@@ -16957,6 +16957,115 @@ class AsOrganisationVoicemailSettingsAPI(AsApiChild, base='telephony/config/voic
         await self.put(url, data=data, params=params)
 
 
+class AsPSTNApi(AsApiChild, base='telephony/pstn/locations'):
+    """
+    PSTN
+
+    PSTN Location Connection Settings supports PSTN selection when creating a location or changing a PSTN type for a
+    location. This is only supported for Local Gateway and Non-integrated CCP.
+
+    Viewing these read-only organization settings requires a full or read-only administrator auth token with a scope of
+    `spark-admin:telephony_pstn_read`.
+
+    Modifying these organization settings requires a full administrator auth token with a scope of
+    `spark-admin:telephony_pstn_write`.
+
+    A partner administrator can retrieve or change settings in a customer's organization using the optional `orgId`
+    query parameter.
+    """
+
+    async def list(self, location_id: str,
+             org_id: str = None) -> list[PSTNConnectionOption]:
+        """
+        Retrieve PSTN Connection Options for a Location
+
+        Retrieve the list of PSTN connection options available for a location.
+
+        PSTN location connection settings enables the admin to configure or change the PSTN provider for a location.
+
+        Retrieving this list requires a full or read-only administrator auth token with a scope of
+        `spark-admin:telephony_pstn_read`.
+
+        :param location_id: Return the list of List PSTN location connection options for this location.
+        :type location_id: str
+        :param org_id: List PSTN location connection options for this organization.
+        :type org_id: str
+        :rtype: list[PSTNConnectionOption]
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        url = self.ep(f'{location_id}/connectionOptions')
+        data = await super().get(url, params=params)
+        r = TypeAdapter(list[PSTNConnectionOption]).validate_python(data['items'])
+        return r
+
+    async def configure(self, location_id: str, id: str = None, premise_route_type: str = None,
+                  premise_route_id: str = None, org_id: str = None):
+        """
+        Setup PSTN Connection for a Location
+
+        Set up or update the PSTN connection details for a location.
+
+        PSTN location connection settings enables the admin to configure or change the PSTN provider for a location.
+
+        Setting up PSTN connection on a location requires a full administrator auth token with a scope of
+        `spark-admin:telephony_pstn_write`.
+
+        :param location_id: Setup PSTN location connection options for this location.
+        :type location_id: str
+        :param id: A unique identifier for the connection. This is required for non-integrated CCP.
+        :type id: str
+        :param premise_route_type: Premise route type. The possible types are TRUNK and ROUTE_GROUP. This is required
+            for the local gateway.
+        :type premise_route_type: str
+        :param premise_route_id: Premise route ID. This refers to either a Trunk ID or a Route Group ID and is required
+            for the local gateway.
+        :type premise_route_id: str
+        :param org_id: Setup PSTN location connection for this organization.
+        :type org_id: str
+        :rtype: None
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        body = dict()
+        if id is not None:
+            body['id'] = id
+        if premise_route_type is not None:
+            body['premiseRouteType'] = premise_route_type
+        if premise_route_id is not None:
+            body['premiseRouteId'] = premise_route_id
+        url = self.ep(f'{location_id}/connection')
+        await super().put(url, params=params, json=body)
+
+    async def read(self, location_id: str,
+             org_id: str = None) -> PSTNConnectionOption:
+        """
+        Retrieve PSTN Connection for a Location
+
+        Retrieves the current configured PSTN connection details for a location.
+
+        PSTN location connection settings enables the admin to configure or change the PSTN provider for a location.
+
+        Retrieving the PSTN connection details for a location requires a full or read-only administrator auth token
+        with a scope of `spark-admin:telephony_pstn_read`.
+
+        :param location_id: Retrieve PSTN location connection details for this location.
+        :type location_id: str
+        :param org_id: Retrieve PSTN location connection details for this organization.
+        :type org_id: str
+        :rtype: :class:`PSTNConnectionOption`
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        url = self.ep(f'{location_id}/connection')
+        data = await super().get(url, params=params)
+        r = PSTNConnectionOption.model_validate(data)
+        return r
+
+
 class AsPagingApi(AsApiChild, base='telephony/config'):
 
     def _endpoint(self, *, location_id: str = None, paging_id: str = None) -> str:
@@ -21459,8 +21568,9 @@ class AsTelephonyApi(AsApiChild, base='telephony/config'):
     permissions_out: AsOutgoingPermissionsApi
     pickup: AsCallPickupApi
     playlist: AsPlayListApi
-    prem_pstn: AsPremisePstnApi
     pnc: AsPrivateNetworkConnectApi
+    prem_pstn: AsPremisePstnApi
+    pstn: AsPSTNApi
     schedules: AsScheduleApi
     supervisors: AsSupervisorApi
     virtual_lines: AsVirtualLinesApi
@@ -21497,6 +21607,7 @@ class AsTelephonyApi(AsApiChild, base='telephony/config'):
         self.playlist = AsPlayListApi(session=session)
         self.pnc = AsPrivateNetworkConnectApi(session=session)
         self.prem_pstn = AsPremisePstnApi(session=session)
+        self.pstn = AsPSTNApi(session=session)
         self.schedules = AsScheduleApi(session=session, base=ScheduleApiBase.locations)
         self.supervisors = AsSupervisorApi(session=session)
         self.virtual_lines = AsVirtualLinesApi(session=session)